Make sure we use the force reconcile method if needed and we dump the state if an upgrade is stuck

+++ e2e/fixtures/ha_fdb_cluster.go
@@ -247,7 +247,7 @@ func (haFDBCluster *HaFdbCluster) UpgradeCluster(version string, waitForReconcil
 func (haFDBCluster *HaFdbCluster) UpgradeClusterWithTimeout(
 	version string,
 	waitForReconciliation bool,
-	timeout int64,
+	timeout int,
 ) error {
 	expectedGenerations := map[string]int64{}
 	for _, cluster := range haFDBCluster.GetAllClusters() {
